flannel: extraRoutes delegate option
Routes listed under "extraRoutes" element of the "delegate"
configuration are appended to the basic route to the pods
network when invoking the delegate plugin.

  {
    "name": "cbr0",
    "cniVersion": "0.3.1",
    "type": "flannel",
    "delegate": {
      "hairpinMode": true,
      "isDefaultGateway": false,
      "extraRoutes": [
        {
          "dst": "192.168.242.0/24"
        },
        {
          "dst": "10.96.0.0/12"
        }
      ]
    }
    ...
  }

This is expecially useful to add back the route to the
cluster services and/or to the hosts (HostPort) when using
isDefaultGateway=false (no default route => only pod-to-pod
connectivity).

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>